move `getStateFromZipCode` helper
