+++ next.config.js
@@ -55,7 +55,7 @@ module.exports = (phase, { defaultConfig }) => {
     env: {
       OVERRIDE_NODE_ENV: OVERRIDE_NODE_ENV || '',
       FB_PIXEL_ID,
-      HIGHSCHOOL_FORM_ACTIVE: isHsFormActive,
+      HIGHSCHOOL_FORM_ACTIVE_UNTIL,
       HIGHSCHOOL_FORM_RESPONSES_NAME,
     },
