+++ src/Forms/Form.Workforce.tsx
@@ -11,9 +11,9 @@ import { MdOpenInNew as NewTabIcon } from 'react-icons/md';
 import Button from '@this/components/Elements/Button';
 import Spinner from '@this/components/Elements/Spinner';
 import { Form, Input, useForm } from '@this/components/Form';
-import { getStateFromZipCode } from '@this/components/Form/helpers';
 import { TOption } from '@this/data/types/bits';
 import { IInfoSessionFormValues } from '@this/data/types/infoSession';
+import { getStateFromZipCode } from '@this/helpers/zipLookup';
 import useKeyCombo from '@this/hooks/useKeyCombo';
 import { pixel } from '@this/lib/pixel';
 import { ISessionDates } from '@this/pages-api/infoSession/dates';
